Embed UUID and SERVER_URL in build types
UUID and SERVER_URL is now part of the build process.
For debug, the placeholder default values are set, however if corresponding env values are set, it sets them.
So by default the settings now displays the environment URL and UUID fetched during build.
For Release, the env values are pulled again from env variable. Which is injected in github action.